finished reordering of file structure
