+++ backend/logic/simulation_2de.py
@@ -1,0 +1,178 @@
+from io import StringIO
+import math
+import re
+from typing import Any, Dict, List
+from fastapi import File, UploadFile
+import numpy as np
+from Bio import SeqIO
+from protein_util import Protein
+
+
+class Electro2d_util():
+    def simulate_ief(proteins, ph_range, canvas_width, canvas_height, steps=25):
+        min_ph = ph_range['min']
+        max_ph = ph_range['max']
+        simulation_results = []
+
+        for step in range(steps + 1):
+            progress = step / steps
+            step_results = []
+            for protein in proteins:
+                protein_data = protein.copy()
+                clampedPH = min(max(protein['pH'], min_ph), max_ph)
+                targetX = Protein.get_ph_position(clampedPH, canvas_width, min_ph, max_ph)
+
+                if step == 0:
+                    startX = np.random.uniform(50, canvas_width - 50)
+                    spreadY = np.random.uniform(50, 70)
+                    protein_data.update({
+                        'x': startX,
+                        'y': spreadY,
+                        'currentpH': min_ph + ((startX - 50) / (canvas_width - 100)) * (max_ph - min_ph),
+                        'bandWidth': 40,
+                        'settled': False
+                    })
+                else:
+                    prev_data = simulation_results[step - 1][proteins.index(protein)]
+                    dx = targetX - prev_data['x']
+                    newX = prev_data['x'] + dx * (0.1 + progress * 0.2)
+                    newBandWidth = max(3, prev_data['bandWidth'] * (1 - progress * 0.8))
+                    settled = abs(dx) < 1
+                    protein_data.update({
+                        'x': newX,
+                        'y': 80,
+                        'bandWidth': newBandWidth,
+                        'settled': settled
+                    })
+                step_results.append(protein_data)
+            simulation_results.append(step_results)
+
+        return simulation_results
+
+
+    def simulate_sds(proteins, y_axis_mode, acrylamide_percentage, canvas_height, steps=25):
+        simulation_results = []
+        condensed_proteins = []
+        for protein in proteins:
+            protein_data = protein.copy()
+            protein_data.update({
+                'y': 150,
+                'condensing': True,
+                'bandWidth': 3
+            })
+            condensed_proteins.append(protein_data)
+        simulation_results.append(condensed_proteins)
+
+        for step in range(1, steps + 1):
+            step_results = []
+            for i, protein in enumerate(proteins):
+                protein_data = protein.copy()
+                prev_data = simulation_results[step - 1][i]
+                if y_axis_mode == 'mw':
+                    targetPosY = Protein.get_mw_position(protein['mw'], canvas_height, acrylamide_percentage)
+                else:
+                    targetPosY = Protein.get_distance_position(protein['mw'], canvas_height, acrylamide_percentage)
+
+                if targetPosY >= 600:
+                    targetPosY = 600
+
+                protein_data.update({
+                    'x': prev_data['x'],
+                    'y': prev_data['y'] + (targetPosY - prev_data['y']) * 0.1,
+                    'condensing': False,
+                    'bandWidth': prev_data['bandWidth']
+                })
+                step_results.append(protein_data)
+            simulation_results.append(step_results)
+
+        return simulation_results
+    def parse_fasta(files: List[UploadFile] = File(...)):
+
+        color_palette = [
+            '#FF0000', '#00FF00', '#0000FF', '#FFFF00', '#FF00FF', '#00FFFF',
+            '#FFA500', '#800080', '#008000', '#FFC0CB', '#A52A2A', '#808080'
+        ]
+        new_proteins = []
+
+        for i, file in enumerate(files):
+            content = file.read()
+            sequences = Protein.parse_fasta_content(content.decode("utf-8"))
+
+            # Collect all IDs from this file
+            id_list = [seq['header'].split("|")[1] for seq in sequences]
+
+            # Fetch links for all IDs in this file
+            links_dict = Protein.find_links(id_list)
+
+            for seq in sequences:
+                pid = seq['header'].split("|")[1]
+                display_name = seq['header'].split("|")[-1]
+
+                # Extract UniProt ID if present
+                uniprot_match = re.search(
+                    r'[OPQ][0-9][A-Z0-9]{3}[0-9]|[A-NR-Z][0-9]([A-Z][A-Z0-9]{2}[0-9]){1,2}',
+                    seq['header']
+                )
+                uniprotId = uniprot_match.group(0) if uniprot_match else "N/A"
+
+                protein_info = {}
+                protein_info['name'] = seq['name']
+                protein_info['fullName'] = seq['name']
+                protein_info['organism'] = seq['organism']
+                protein_info['uniprotId'] = uniprotId
+                protein_info['mw'] = seq['mw']
+                protein_info['pH'] = seq['pH']
+                protein_info['color'] = color_palette[len(new_proteins) % len(color_palette)]
+                protein_info['sequence'] = seq['sequence']
+                protein_info['x'] = 50
+                protein_info['y'] = 300
+                protein_info['currentpH'] = 7
+                protein_info['velocity'] = 0
+                protein_info['settled'] = False
+                protein_info['ID'] = pid
+                protein_info['Link'] = links_dict.get(pid) or "N/A"
+                protein_info['display_name'] = display_name
+
+                new_proteins.append(protein_info.copy())
+
+        return new_proteins
+    def parse_fasta_content(content: str) -> List[Dict[str, Any]]:
+        sequences = []
+        fasta_io = StringIO(content)
+        for record in SeqIO.parse(fasta_io, "fasta"):
+            header = str(record.description)
+            sequence = str(record.seq)
+            mw = Protein.calculate_molecular_weight(sequence)
+            pH = Protein.calculate_theoretical_pi(sequence)
+            
+            info = Protein.extract_protein_info(header)
+            sequences.append({
+                'header': header,
+                'sequence': sequence,
+                'name': info['name'],
+                'organism': info['organism'],
+                'mw': mw,
+                'pH': pH
+            })
+        return sequences
+    
+    def get_ph_position(pH, canvas_width, min_ph, max_ph):
+        clampedPH = min(max(pH, min_ph), max_ph)
+        return 50 + ((clampedPH - min_ph) / (max_ph - min_ph)) * (canvas_width - 100)
+
+
+    def get_mw_position(mw, canvas_height, acrylamide_percentage):
+        min_mw = 1000
+        max_mw = 1000000
+        log_mw = math.log10(min(max(mw, min_mw), max_mw))
+        acrylamide_factor = 1 + (acrylamide_percentage - 7.5) / 15
+        return 170 + ((math.log10(max_mw) - log_mw) / (math.log10(max_mw) - math.log10(min_mw))) * (canvas_height - 220) * acrylamide_factor
+
+
+    def get_distance_position(mw, canvas_height, acrylamide_percentage, max_distance_traveled=6):
+        min_mw = 1000
+        max_mw = 1000000
+        normalized_mw = (math.log10(min(max(mw, min_mw), max_mw)) - math.log10(min_mw)) / (math.log10(max_mw) - math.log10(min_mw))
+        acrylamide_factor = 1 + (acrylamide_percentage - 7.5) / 10
+        distance = max_distance_traveled * (1 - normalized_mw) * acrylamide_factor
+        return 170 + (distance / (max_distance_traveled * acrylamide_factor)) * (canvas_height - 220)
