+++ backend/server.py
@@ -2,7 +2,8 @@ import uvicorn
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
-import root
+import backend.api.one_de_routes as one_de_routes
+import backend.api.two_de_routes as two_de_routes
 
 
 
@@ -49,7 +50,8 @@ app.add_middleware(
 )
 
 # Routers
-app.include_router(root.router)
+app.include_router(one_de_routes.router)
+app.include_router(two_de_routes.router)
 
 ''' 
 NOTE FOR FUTURE DEVELOPERS:
